les4/django_server/Django-Server2-main/requirements.txt
asgiref==3.5.2
Django==4.0.4
gunicorn==20.1.0
sqlparse==0.4.2
django-cors-headers==3.13.0
djangorestframework==3.13.1
openpyxl==3.0.10
pandas==1.4.3
gunicorn==20.1.0